fix(nav) Replace dropdown with account link
