+++ src/components/Header/Header.tsx
@@ -1,7 +1,5 @@
-import { useCallback } from "react";
-import { NavLink, useNavigate } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 import { ACCOUNT_PATH } from "services/constants/routes";
-import { supabase } from "services/supabase";
 
 import logo from "./logo.webp";
 
@@ -10,12 +8,6 @@ type HeaderProps = {
 };
 
 function Header({ isAuthenticated }: HeaderProps) {
-  const navigate = useNavigate();
-  const onLogoutClick = useCallback(async () => {
-    await supabase.auth.signOut();
-    navigate("/");
-  }, [navigate]);
-
   return (
     <div className="navbar bg-base-100 sm:mb-16">
       <div className="flex-1 text-primary">
@@ -24,39 +16,17 @@ function Header({ isAuthenticated }: HeaderProps) {
         </NavLink>
       </div>
       <div className="flex-none gap-2">
-        {!isAuthenticated ? (
-          <NavLink data-testid="loginButton" className="btn" to="/signin">
-            Get Started
-          </NavLink>
-        ) : null}
-        {isAuthenticated ? (
-          <div className="dropdown dropdown-end">
-            <button tabIndex={0} className="btn btn-ghost" type="button">
-              My Account
-            </button>
-            <ul
-              tabIndex={0}
-              role="menu"
-              className="mt-3 p-2 shadow menu menu-compact dropdown-content bg-base-100 rounded-box w-52"
-            >
-              <li>
-                <NavLink className="justify-between" to={ACCOUNT_PATH}>
-                  Settings
-                </NavLink>
-              </li>
-              <li>
-                <div
-                  role="button"
-                  onClick={onLogoutClick}
-                  onKeyDown={onLogoutClick}
-                  tabIndex={0}
-                >
-                  Logout
-                </div>
-              </li>
-            </ul>
-          </div>
-        ) : null}
+        <ul className="menu menu-horizontal p-0">
+          <li>
+            {!isAuthenticated ? (
+              <NavLink data-testid="loginButton" to="/signin">
+                Get Started
+              </NavLink>
+            ) : (
+              <NavLink to={ACCOUNT_PATH}>My Account</NavLink>
+            )}
+          </li>
+        </ul>
       </div>
     </div>
   );
